Stub out expand_tsv, and replace README image
