+++ script-docs/README-expand_tsv.py.md
@@ -1,0 +1,11 @@
+```
+usage: expand_tsv.py [-h] [--path PATH]
+
+Given an assay TSV, determines its type, and for each field with an ontology
+mapping, adds an additional column with the URI that corresponds to the
+supplied term, and prints the result to STDOUT.
+
+optional arguments:
+  -h, --help   show this help message and exit
+  --path PATH
+```
