--- script-docs/README-lookup_ontology_uris.py.md
@@ -1,10 +1,0 @@
-```
-usage: lookup_ontology_uris.py [-h]
-
-Given an assay TSV on STDIN, determines its type, and for each field with an
-ontology mapping, adds an additional column with the URI that corresponds to
-the supplied term, and prints the result to STDOUT.
-
-optional arguments:
-  -h, --help  show this help message and exit
-```
